refactor: Move token card to extra component
